change define in Bot()
