--- core/classes.py
@@ -1,18 +1,0 @@
-# -*- coding: UTF-8 -*-
-import discord
-from discord.ext import commands
-
-class Cogs(commands.Cog):
-    def __init__(self, bot):
-        self.bot=bot
-
-class Slash_Cogs(commands.Cog):
-    def __init__(self,bot):
-        self.bot=bot
-        self.count = 0
-        
-
-
-
-
-
